requirements.txt
# Core
langchain==0.3.15
langchain-openai==0.2.14
openai==1.58.1

# Interface
streamlit==1.40.2

# External tools / HTTP
requests==2.31.0
python-dotenv==1.0.0